ASYLUM: rename detection_table.h
All engines use the plural "detection_tables.h"
