+++ engines/asylum/detection_tables.h
